LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-self-baseline-not-applied-if-sizing-cyclic-dependency-003-expected.txt

flex rows - column-axis baseline - the blue orthogonal item didn't participate in the first iteration
É
É
É
 flex row - column-axis baseline - the blue relative sized item didn't participate in the first iterarion
É
É
 flex row - both the blue relative sized and red orthogonal didn't participate in the first iteration
É
É
 flex column - row-axis baseline - the blue orthogonal item didn't participate in the first iteration
É
É
É
 flex column - column-axis baseline - the blue relative sized item didn't participate in the first iterarion
É
É
 flex columns - both the blue relative sized and red orthogonal didn't participate in the first iteration
É
É

FAIL .grid 1 assert_equals: 
<div class="grid column justifyItemsBaseline">
  <div class="item1 relativeWidth height50" data-offset-x="16" data-offset-y="0" data-expected-width="40" data-expected-height="50"></div>
  <div class="item2" data-offset-x="16" data-offset-y="50" data-expected-width="20" data-expected-height="50">É</div>
  <div class="item3 verticalLR" data-offset-x="0" data-offset-y="100" data-expected-width="80" data-expected-height="100">É</div>
</div>
offsetLeft expected 16 but got 0
FAIL .grid 2 assert_equals: 
<div class="grid column justifyItemsBaseline">
  <div class="item1 relativeWidth height50" data-offset-x="16" data-offset-y="0" data-expected-width="40" data-expected-height="50"></div>
  <div class="item2 verticalLR" data-offset-x="12" data-offset-y="50" data-expected-width="20" data-expected-height="50">É</div>
  <div class="item3 verticalLR" data-offset-x="0" data-offset-y="100" data-expected-width="80" data-expected-height="100">É</div>
</div>
offsetLeft expected 16 but got 0
FAIL .grid 3 assert_equals: 
<div class="grid column justifyItemsBaseline">
  <div class="item1" data-offset-x="16" data-offset-y="0" data-expected-width="30" data-expected-height="50">É</div>
  <div class="item2 verticalLR" data-offset-x="12" data-offset-y="50" data-expected-width="20" data-expected-height="50">É</div>
  <div class="item3 verticalLR" data-offset-x="0" data-offset-y="100" data-expected-width="80" data-expected-height="100">É</div>
</div>
offsetLeft expected 16 but got 0
FAIL .grid 4 assert_equals: 
<div class="grid row alignItemsBaseline">
  <div class="item1 relativeHeight" data-offset-x="0" data-offset-y="24" data-expected-width="50" data-expected-height="40"></div>
  <div class="item2 verticalLR" data-offset-x="50" data-offset-y="44" data-expected-width="50" data-expected-height="20">É</div>
  <div class="item3" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="80">É</div>
</div>
offsetTop expected 24 but got 0
FAIL .grid 5 assert_equals: 
<div class="grid row alignItemsBaseline ">
  <div class="item1 relativeHeight" data-offset-x="0" data-offset-y="40" data-expected-width="50" data-expected-height="40"></div>
  <div class="item2" data-offset-x="50" data-offset-y="64" data-expected-width="50" data-expected-height="20">É</div>
  <div class="item3 verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="80">É</div>
</div>
offsetTop expected 40 but got 0
FAIL .grid 6 assert_equals: 
<div class="grid row alignItemsBaseline">
  <div class="item1 verticalLR" data-offset-x="0" data-offset-y="34" data-expected-width="50" data-expected-height="30">É</div>
  <div class="item2" data-offset-x="50" data-offset-y="48" data-expected-width="50" data-expected-height="20">É</div>
  <div class="item3" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="80">É</div>
</div>
offsetTop expected 34 but got 0
